Fix dumb warnings

+++ src/Joystick.cpp
@@ -7,7 +7,7 @@ std::vector<Joystick> Joystick::Joysticks()
 	std::vector<Joystick> out;
 	for(size_t i=0;i<GLFW_JOYSTICK_LAST+1;i++)
 	{
-		out.push_back(i);
+		out.push_back((unsigned int)i);
 	}
 	return out;
 }
